+++ tests/Rule/LowerCaseRuleTest.php
@@ -9,21 +9,21 @@ use PHPUnit\Framework\TestCase;
 use Stadly\PasswordPolice\ValidationError;
 
 /**
- * @coversDefaultClass \Stadly\PasswordPolice\Rule\LowerCase
+ * @coversDefaultClass \Stadly\PasswordPolice\Rule\LowerCaseRule
  * @covers ::<protected>
  * @covers ::<private>
  */
-final class LowerCaseTest extends TestCase
+final class LowerCaseRuleTest extends TestCase
 {
     /**
      * @covers ::__construct
      */
     public function testCanConstructRuleWithMinConstraint(): void
     {
-        $rule = new LowerCase(5, null);
+        $rule = new LowerCaseRule(5, null);
 
         // Force generation of code coverage
-        $ruleConstruct = new LowerCase(5, null);
+        $ruleConstruct = new LowerCaseRule(5, null);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -32,10 +32,10 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanConstructRuleWithMaxConstraint(): void
     {
-        $rule = new LowerCase(0, 10);
+        $rule = new LowerCaseRule(0, 10);
 
         // Force generation of code coverage
-        $ruleConstruct = new LowerCase(0, 10);
+        $ruleConstruct = new LowerCaseRule(0, 10);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -44,10 +44,10 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanConstructRuleWithBothMinAndMaxConstraint(): void
     {
-        $rule = new LowerCase(5, 10);
+        $rule = new LowerCaseRule(5, 10);
 
         // Force generation of code coverage
-        $ruleConstruct = new LowerCase(5, 10);
+        $ruleConstruct = new LowerCaseRule(5, 10);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -58,7 +58,7 @@ final class LowerCaseTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $rule = new LowerCase(-10, null);
+        $rule = new LowerCaseRule(-10, null);
     }
 
     /**
@@ -68,7 +68,7 @@ final class LowerCaseTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $rule = new LowerCase(10, 5);
+        $rule = new LowerCaseRule(10, 5);
     }
 
     /**
@@ -76,10 +76,10 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanConstructUnconstrainedRule(): void
     {
-        $rule = new LowerCase(0, null);
+        $rule = new LowerCaseRule(0, null);
 
         // Force generation of code coverage
-        $ruleConstruct = new LowerCase(0, null);
+        $ruleConstruct = new LowerCaseRule(0, null);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -88,10 +88,10 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanConstructRuleWithMinConstraintEqualToMaxConstraint(): void
     {
-        $rule = new LowerCase(5, 5);
+        $rule = new LowerCaseRule(5, 5);
 
         // Force generation of code coverage
-        $ruleConstruct = new LowerCase(5, 5);
+        $ruleConstruct = new LowerCaseRule(5, 5);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -100,11 +100,11 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanAddConstraint(): void
     {
-        $rule = new LowerCase(5, 5, 1);
+        $rule = new LowerCaseRule(5, 5, 1);
         $rule->addConstraint(10, 10, 1);
 
         // Force generation of code coverage
-        $ruleConstruct = new LowerCase(5, 5, 1);
+        $ruleConstruct = new LowerCaseRule(5, 5, 1);
         $ruleConstruct->addConstraint(10, 10, 1);
         self::assertEquals($rule, $ruleConstruct);
     }
@@ -114,10 +114,10 @@ final class LowerCaseTest extends TestCase
      */
     public function testConstraintsAreOrdered(): void
     {
-        $rule = new LowerCase(5, 5, 1);
+        $rule = new LowerCaseRule(5, 5, 1);
         $rule->addConstraint(10, 10, 2);
 
-        $ruleConstruct = new LowerCase(10, 10, 2);
+        $ruleConstruct = new LowerCaseRule(10, 10, 2);
         $ruleConstruct->addConstraint(5, 5, 1);
         self::assertEquals($rule, $ruleConstruct);
     }
@@ -127,7 +127,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testMinConstraintCanBeSatisfied(): void
     {
-        $rule = new LowerCase(2, null);
+        $rule = new LowerCaseRule(2, null);
 
         self::assertTrue($rule->test('FOO bar'));
     }
@@ -137,7 +137,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testMinConstraintCanBeUnsatisfied(): void
     {
-        $rule = new LowerCase(2, null);
+        $rule = new LowerCaseRule(2, null);
 
         self::assertFalse($rule->test('FOO BAR'));
     }
@@ -147,7 +147,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testMaxConstraintCanBeSatisfied(): void
     {
-        $rule = new LowerCase(0, 3);
+        $rule = new LowerCaseRule(0, 3);
 
         self::assertTrue($rule->test('FOO bar'));
     }
@@ -157,7 +157,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testMaxConstraintCanBeUnsatisfied(): void
     {
-        $rule = new LowerCase(0, 3);
+        $rule = new LowerCaseRule(0, 3);
 
         self::assertFalse($rule->test('foo bar'));
     }
@@ -167,7 +167,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testLowerCaseUtf8IsCounted(): void
     {
-        $rule = new LowerCase(1, null);
+        $rule = new LowerCaseRule(1, null);
 
         self::assertTrue($rule->test('á'));
     }
@@ -177,7 +177,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testUpperCaseUtf8IsNotCounted(): void
     {
-        $rule = new LowerCase(1, null);
+        $rule = new LowerCaseRule(1, null);
 
         self::assertFalse($rule->test('Á'));
     }
@@ -187,7 +187,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testRuleIsSatisfiedWhenConstraintWeightIsLowerThanTestWeight(): void
     {
-        $rule = new LowerCase(0, 3, 1);
+        $rule = new LowerCaseRule(0, 3, 1);
 
         self::assertTrue($rule->test('foo bar', 2));
     }
@@ -197,7 +197,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testRuleCanBeValidated(): void
     {
-        $rule = new LowerCase(1, null);
+        $rule = new LowerCaseRule(1, null);
 
         self::assertNull($rule->validate('foo'));
     }
@@ -207,7 +207,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testRuleWithMinConstraintCanBeInvalidated(): void
     {
-        $rule = new LowerCase(5, null);
+        $rule = new LowerCaseRule(5, null);
 
         self::assertEquals(
             new ValidationError('There must be at least 5 lower case characters.', 'Foo', $rule, 1),
@@ -220,7 +220,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testRuleWithMaxConstraintCanBeInvalidated(): void
     {
-        $rule = new LowerCase(0, 10);
+        $rule = new LowerCaseRule(0, 10);
 
         self::assertEquals(
             new ValidationError('There must be at most 10 lower case characters.', 'Foo bar qwerty test', $rule, 1),
@@ -233,7 +233,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testRuleWithBothMinAndMaxConstraintCanBeInvalidated(): void
     {
-        $rule = new LowerCase(5, 10);
+        $rule = new LowerCaseRule(5, 10);
 
         self::assertEquals(
             new ValidationError('There must be between 5 and 10 lower case characters.', 'Foo', $rule, 1),
@@ -246,7 +246,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testRuleWithMaxConstraintEqualToZeroCanBeInvalidated(): void
     {
-        $rule = new LowerCase(0, 0);
+        $rule = new LowerCaseRule(0, 0);
 
         self::assertEquals(
             new ValidationError('There must be no lower case characters.', 'Foo', $rule, 1),
@@ -259,7 +259,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testRuleWithMinConstraintEqualToMaxConstraintCanBeInvalidated(): void
     {
-        $rule = new LowerCase(3, 3);
+        $rule = new LowerCaseRule(3, 3);
 
         self::assertEquals(
             new ValidationError('There must be exactly 3 lower case characters.', 'Foo', $rule, 1),